Refactor. Move prompt logic inside Prompt class. Fix filename conventions. SD 1.5 still broken.
