+++ artspew.py
@@ -3,11 +3,8 @@ import os
 import argparse
 import torch
 import logging
-from PIL import Image, PngImagePlugin
-import piexif
-import piexif.helper
-from src.StableDiffusionSD15 import StableDiffusionSD15
-from src.StableDiffusionSDXL import StableDiffusionSDXL
+from src.stable_diffusion_sd15 import StableDiffusionSD15
+from src.stable_diffusion_sdxl import StableDiffusionSDXL
 from pathvalidate import sanitize_filename
 
 MODEL_ID_SD15 = 'runwayml/stable-diffusion-v1-5'
@@ -68,46 +65,6 @@ def parse_arguments():
     return args
 
 
-def save_image_with_geninfo(image, geninfo, filename, extension=None, existing_pnginfo=None, pnginfo_section_name='parameters'):
-    """
-    Saves image to filename, including geninfo as text information for generation info.
-    For PNG images, geninfo is added to existing pnginfo dictionary using the pnginfo_section_name argument as key.
-    For JPG images, there's no dictionary and geninfo just replaces the EXIF description.
-    """
-
-    if extension is None:
-        extension = os.path.splitext(filename)[1]
-
-    image_format = Image.registered_extensions()[extension]
-
-    if extension.lower() == '.png':
-        existing_pnginfo = existing_pnginfo or {}
-        existing_pnginfo[pnginfo_section_name] = geninfo
-        pnginfo_data = PngImagePlugin.PngInfo()
-        for k, v in (existing_pnginfo or {}).items():
-            pnginfo_data.add_text(k, str(v))
-        image.save(filename, format=image_format, pnginfo=pnginfo_data)
-
-    elif extension.lower() in (".jpg", ".jpeg", ".webp"):
-        if image.mode == 'RGBA':
-            image = image.convert("RGB")
-        elif image.mode == 'I;16':
-            image = image.point(lambda p: p * 0.0038910505836576).convert("RGB" if extension.lower() == ".webp" else "L")
-
-        image.save(filename, format=image_format)
-
-        if geninfo is not None:
-            exif_bytes = piexif.dump({
-                "Exif": {
-                    piexif.ExifIFD.UserComment: piexif.helper.UserComment.dump(geninfo or "", encoding="unicode")
-                },
-            })
-
-            piexif.insert(exif_bytes, filename)
-    else:
-        image.save(filename, format=image_format, extension='.jpg')
-
-
 def main():
     args = parse_arguments()
 
@@ -133,16 +90,15 @@ def main():
     torch.manual_seed(seed)
 
     if args.xl:
-        model = StableDiffusionSDXL(args.model_id, args.tiny_vae, args.lcm, args.width, args.height, args.batch_count, args.batch_size, args.random_tokens, args.steps, args.guidance, args.torch_compile)
+        model = StableDiffusionSDXL(args.model_id, args.tiny_vae, args.lcm, args.width, args.height, seed, args.batch_count, args.batch_size, args.random_tokens, args.steps, args.guidance, args.torch_compile)
     else:
-        model = StableDiffusionSD15(args.model_id, args.tiny_vae, args.lcm, args.width, args.height, args.batch_count, args.batch_size, args.random_tokens, args.steps, args.guidance, args.torch_compile)
+        model = StableDiffusionSD15(args.model_id, args.tiny_vae, args.lcm, args.width, args.height, seed, args.batch_count, args.batch_size, args.random_tokens, args.steps, args.guidance, args.torch_compile)
 
     sequence_number = -1
     if not os.path.exists('spew'):
         os.makedirs('spew')
 
-    files = [entry.name for entry in os.scandir('spew')
-                if entry.name.startswith(model.get_filename_prefix())]
+    files = [entry.name for entry in os.scandir('spew') if entry.name.startswith(model.get_filename_prefix())]
 
     if files:
         sorted_files = sorted(files, key=lambda x: int(x.split('-')[1]))
@@ -151,9 +107,8 @@ def main():
     images = model.generate_images(args.prompt)
     for idx, image in enumerate(images):
         sequence_number += 1
-        geninfo = f"{image['prompt']}\nSteps: {args.steps}, Sampler: Euler a, CFG scale: {args.guidance}, Seed: {seed}, Size: {args.width}x{args.height}, Model: {args.model_id}"
-        safe_prompt = sanitize_filename(image['prompt'])
-        save_image_with_geninfo(image['image'], geninfo, f"spew/{model.get_filename_prefix()}{sequence_number:09d}-{idx:02d}-{safe_prompt}.jpg")
+        safe_prompt = sanitize_filename(image.prompt_text)
+        image.save(f"spew/{model.get_filename_prefix()}{sequence_number:09d}-{idx:02d}-{safe_prompt}.jpg")
 
 
 if __name__ == "__main__":
